+++ CHANGELOG.md
@@ -4,3 +4,6 @@
 [Full git log](https://github.com/peps1/airdcpp-advanced-sharing/compare/5918334e8dec8cfcb9c639583a2f24bc9b6a5aa0...v1.0.0)
 
 Initial release
+
+* Set refresh queue limit and auto stop refresh when that limit is reached
+* /stophash command to stop all current hashers and clear the refresh queue